machine_b/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
astroid==2.2.5
backcall==0.1.0
bcrypt==3.1.7
billiard==3.3.0.23
cached-property==1.5.1
celery==3.1.18
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
decorator==4.4.0
Django==2.2.2
docker==3.7.2
docker-compose==1.24.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
flake8==3.7.7
idna==2.7
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.20
jedi==0.13.3
jsonschema==2.6.0
kombu==3.0.37
lazy-object-proxy==1.4.1
mccabe==0.6.1
paramiko==2.5.0
parso==0.4.0
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3.1
pymongo==3.8.0
PyNaCl==1.3.0
python-decouple==3.1
pytz==2019.1
PyYAML==3.13
redis==2.10.3
requests==2.20.1
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
texttable==0.9.1
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.24.3
wcwidth==0.1.7
websocket-client==0.56.0
wrapt==1.11.1
xlrd==1.2.0
